sac-esame15/backend/requirements.txt
flask
flask-restful
gunicorn
marshmallow
google-cloud-firestore
validators
python-Levenshtein
